Move the get_cycle_hours and check_forecast_consistency methods from reliability calibration to be separate functions that can be used by EMOS.
