+++ improver/calibration/reliability_calibration.py
@@ -35,7 +35,6 @@ import warnings
 import iris
 import numpy as np
 import scipy
-from iris.exceptions import CoordinateNotFoundError
 
 from improver import BasePlugin, PostProcessingPlugin
 from improver.calibration.utilities import (
@@ -179,53 +178,6 @@ class ConstructReliabilityCalibrationTables(BasePlugin):
             self.table_columns, long_name="table_row_name", units=1
         )
         return index_coord, name_coord
-
-    @staticmethod
-    def _get_cycle_hours(forecast_reference_time):
-        """
-        Returns a set of integer representations of the hour of the
-        forecast reference time (the cycle hour).
-
-        Args:
-            forecast_reference_time (iris.coord.DimCoord):
-                The forecast_reference_time coordinate to extract cycle hours
-                from.
-        Returns:
-            set:
-                A set of integer representations of the cycle hours.
-        """
-        cycle_hours = []
-        for frt in forecast_reference_time.cells():
-            cycle_hours.append(np.int32(frt.point.hour))
-        return set(cycle_hours)
-
-    def _check_forecast_consistency(self, forecasts):
-        """
-        Checks that the forecast cubes are all from a consistent cycle and
-        with a consistent forecast period.
-
-        Args:
-            forecasts (iris.cube.Cube):
-        Raises:
-            ValueError: Forecast cubes do not share consistent cycle hour and
-                        forecast period.
-        """
-        n_cycle_hours = len(
-            self._get_cycle_hours(forecasts.coord("forecast_reference_time"))
-        )
-        try:
-            (n_forecast_periods,) = forecasts.coord("forecast_period").shape
-        except CoordinateNotFoundError:
-            n_forecast_periods = 0
-        if n_cycle_hours != 1 or n_forecast_periods != 1:
-            msg = (
-                "Forecasts have been provided from differing cycle hours "
-                "or forecast periods, or without these coordinates. These "
-                "coordinates should be present and consistent between "
-                "forecasts. Number of cycle hours found: {}, number of "
-                "forecast periods found: {}."
-            )
-            raise ValueError(msg.format(n_cycle_hours, n_forecast_periods))
 
     @staticmethod
     def _define_metadata(forecast_slice):
@@ -407,7 +359,7 @@ class ConstructReliabilityCalibrationTables(BasePlugin):
 
         time_coord = historic_forecasts.coord("time")
 
-        self._check_forecast_consistency(historic_forecasts)
+        check_forecast_consistency(historic_forecasts)
         reliability_cube = self._create_reliability_table_cube(
             historic_forecasts, threshold_coord
         )
