+++ improver_tests/calibration/reliability_calibration/test_ConstructReliabilityCalibrationTables.py
@@ -243,94 +243,6 @@ class Test__create_reliability_table_coords(unittest.TestCase):
         assert_array_equal(name_coord.points, expected_names)
 
 
-class Test__get_cycle_hours(Test_Setup):
-
-    """Test the _get_cycle_hours method."""
-
-    def test_single_value(self):
-        """Test that the expected cycle hour value is returned in a set."""
-
-        frt = self.forecast_1.coord("forecast_reference_time")
-        result = Plugin()._get_cycle_hours(frt)
-        self.assertEqual(result, set([0]))
-
-    def test_multiple_values(self):
-        """Test that the expected cycle hour values are returned in a set."""
-
-        frt = self.forecast_1.coord("forecast_reference_time")
-        expected = np.array([0, 1, 4])
-        frt = frt.copy(points=3600 * expected + frt.points[0])
-        result = Plugin()._get_cycle_hours(frt)
-        self.assertEqual(result, set(expected))
-
-
-class Test__check_forecast_consistency(Test_Setup):
-
-    """Test the _check_forecast_consistency method."""
-
-    def test_matching_forecasts(self):
-        """Test case in which forecasts share cycle hour and forecast period
-        values. No result is expected in this case, hence there is no value
-        comparison; the test is the absence of an exception."""
-
-        Plugin()._check_forecast_consistency(self.forecasts)
-
-    def test_unmatching_cycle_hours(self):
-        """Test case in which forecasts do not share consistent cycle hours."""
-
-        self.forecasts.coord("forecast_reference_time").points = [3600, 7200]
-
-        msg = (
-            "Forecasts have been provided from differing cycle hours "
-            "or forecast periods, or without these coordinates. These "
-            "coordinates should be present and consistent between "
-            "forecasts. Number of cycle hours found: 2, number of "
-            "forecast periods found: 1."
-        )
-
-        with self.assertRaisesRegex(ValueError, msg):
-            Plugin()._check_forecast_consistency(self.forecasts)
-
-    def test_unmatching_forecast_periods(self):
-        """Test case in which forecasts do not share consistent forecast
-        periods."""
-
-        forecasts = iris.cube.CubeList()
-        forecast = self.forecasts[0].copy()
-        forecast.coord("forecast_period").points = [3600]
-        forecasts.append(forecast)
-        forecasts.append(self.forecasts[1])
-        forecasts = MergeCubes()(forecasts)
-
-        msg = (
-            "Forecasts have been provided from differing cycle hours "
-            "or forecast periods, or without these coordinates. These "
-            "coordinates should be present and consistent between "
-            "forecasts. Number of cycle hours found: 1, number of "
-            "forecast periods found: 2."
-        )
-
-        with self.assertRaisesRegex(ValueError, msg):
-            Plugin()._check_forecast_consistency(forecasts)
-
-    def test_missing_forecast_period(self):
-        """Test case in which forecasts do not have a forecast period
-        coordinate."""
-
-        self.forecasts.remove_coord("forecast_period")
-
-        msg = (
-            "Forecasts have been provided from differing cycle hours "
-            "or forecast periods, or without these coordinates. These "
-            "coordinates should be present and consistent between "
-            "forecasts. Number of cycle hours found: 1, number of "
-            "forecast periods found: 0."
-        )
-
-        with self.assertRaisesRegex(ValueError, msg):
-            Plugin()._check_forecast_consistency(self.forecasts)
-
-
 class Test__define_metadata(Test_Setup):
 
     """Test the _define_metadata method."""
